Support streaming response
This changed adds support for streaming responses by adding a process
filter. The process filter is based on the default default process
filter as described in the manual. It inserts the response into the
process buffer. Whenever a chunk of the response arrives the THEN
callback is called with this chunk.