Support streaming response
This changed adds support for streaming responses by adding a process
filter and a callback function. The process filter is based on the
default default process filter as described in the manual. It inserts
the response into the process buffer. Additionally, if a callback
function is provided by the `during` keyword, it is called every time
a part of the response body is received.