Add an option to set the process filter
The process filter is an optional function to be used as the process
filter for the curl process.  It can be used to handle HTTP responses
in a streaming way.  The function must accept 2 arguments, the process
object running curl, and a string which is output received from the
process.  The default process filter inserts the output of the process
into the process buffer.  The provided function should at least insert
output up to the HTTP body into the process buffer.